Add camera projection functions

+++ src/input.h
@@ -1,7 +1,7 @@
 #ifndef INPUT_H
 #define INPUT_H
 
-#include <SDL2/SDL.h>
+#include "SDL3/SDL.h"
 
 #define MAX_EVENT_CALLBACKS 64
 
@@ -19,8 +19,8 @@ void input_register(SDL_Keycode key, void (*fn)(bool));
 void input_deregister(SDL_Keycode key, void (*fn)(bool));
 void input_reset(void);
 
-extern int mouse_x;
-extern int mouse_y;
+extern float mouse_x;
+extern float mouse_y;
 
 #endif
